requirements.txt
aiohttp~=3.9.3
aiosqlite~=0.19.0
discord.py
python-dotenv~=1.0.1
requests~=2.31.0
ossapi~=3.4.3
markovify~=0.9.4
spacy~=3.7.4